FileAPI/url/url_createobjecturl_file-manual.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>FileAPI Test: url</title>
    <link rel="author" title="Intel" href="www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/FileAPI/#creating-revoking">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <form name="uploadData">
      <input type="file" id="fileChooser"></input>
    </form>

    <div>
      <p>Test step:</p>
      <ol>
        <li>Download <a href="test.png">test.png</a> to local.</li>
        <li>Select the local file (test.png) to run the test.</li>
      </ol>
    </div>

    <div id="log"></div>

    <script>
    async_test(function () {
      var fileInput = document.querySelector('#fileChooser');
      fileInput.change = this.step_func(function (evt) {
        var file = fileInput.files[0];
        var testBlob = window.URL.createObjectURL(file)

        test(function() {
          assert_true(file instanceof File, "FileList returns File");
        }, "Check if FileList returns File");

        test(function() {
          assert_not_equals(testBlob, null, "createObjectURL returns not-null when parameter is File");
        }, "Check if createObjectURL returns not-null when parameter is File");

        test(function() {
          assert_equals(typeof testBlob, "string", "Blob URI is typeof string");
          assert_equals(testBlob.indexOf("blob"), 0, "Blob URI starts with 'blob'");
        }, "Check if Blob URI returned by createObjectURL starts with 'blob' when parameter is File");

        var testBlob = window.URL.createFor(file);

        test(function() {
          assert_not_equals(testBlob, null, "createObjectURL returns not-null when parameter is File");
        }, "Check if createObjectURL returns not-null when parameter is File and autoRevoke is false");

        test(function() {
          assert_equals(typeof testBlob, "string", "Blob URI is typeof string");
          assert_equals(testBlob.indexOf("blob"), 0, "Blob URI starts with 'blob'");
        }, "Check if Blob URI returned by createObjectURL starts with 'blob' when parameter is File and autoRevoke is false");

        this.done();
      });
    });
    </script>
  </body>
</html>
